[13/n] More -Wconversion fixes

+++ src/video/DummyRenderer.cc
@@ -23,16 +23,16 @@ void DummyRenderer::updateSuperimposing(const RawFrame* /*videoSource*/,
                                         EmuTime::param /*time*/) {
 }
 
-void DummyRenderer::updateForegroundColor(int /*color*/, EmuTime::param /*time*/) {
+void DummyRenderer::updateForegroundColor(byte /*color*/, EmuTime::param /*time*/) {
 }
 
-void DummyRenderer::updateBackgroundColor(int /*color*/, EmuTime::param /*time*/) {
+void DummyRenderer::updateBackgroundColor(byte /*color*/, EmuTime::param /*time*/) {
 }
 
-void DummyRenderer::updateBlinkForegroundColor(int /*color*/, EmuTime::param /*time*/) {
+void DummyRenderer::updateBlinkForegroundColor(byte /*color*/, EmuTime::param /*time*/) {
 }
 
-void DummyRenderer::updateBlinkBackgroundColor(int /*color*/, EmuTime::param /*time*/) {
+void DummyRenderer::updateBlinkBackgroundColor(byte /*color*/, EmuTime::param /*time*/) {
 }
 
 void DummyRenderer::updateBlinkState(bool /*enabled*/, EmuTime::param /*time*/) {
